vertigo/functions/status-generator/requirements.txt
google-cloud-firestore
google-auth
google-generativeai==0.8.5
google-api-python-client
functions-framework
langfuse
google-cloud-secret-manager
